+++ src/main/java/com/brand/blockus/blocks/base/HorizontalAxisBlock.java
@@ -8,10 +8,10 @@ import net.minecraft.state.property.EnumProperty;
 import net.minecraft.state.property.Properties;
 import net.minecraft.util.math.Direction;
 
-public class HorizontalAxisBlockBase extends Block {
+public class HorizontalAxisBlock extends Block {
     public static final EnumProperty<Direction.Axis> HORIZONTAL_AXIS;
 
-    public HorizontalAxisBlockBase(Settings settings) {
+    public HorizontalAxisBlock(Settings settings) {
         super(settings);
     }