+++ src/main/java/com/brand/blockus/blocks/base/WeightStorageCubeBlock.java
@@ -8,8 +8,8 @@ import net.minecraft.item.ItemStack;
 import net.minecraft.item.Items;
 import net.minecraft.sound.SoundCategory;
 import net.minecraft.sound.SoundEvents;
-import net.minecraft.util.ActionResult;
 import net.minecraft.util.Hand;
+import net.minecraft.util.ItemActionResult;
 import net.minecraft.util.hit.BlockHitResult;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.world.World;
@@ -19,22 +19,16 @@ public class WeightStorageCubeBlock extends SimpleFallingBlock {
         super(settings);
     }
 
-    public ActionResult onUse(BlockState state, World world, BlockPos pos, PlayerEntity player, Hand hand, BlockHitResult hit) {
+    public ItemActionResult onUseWithItem(ItemStack stack, BlockState state, World world, BlockPos pos, PlayerEntity player, Hand hand, BlockHitResult hit) {
         ItemStack itemStack = player.getStackInHand(hand);
-        if (!isPoppy(itemStack)) {
-            return ActionResult.PASS;
-        } else if (isPoppy(itemStack)) {
+        if (stack.isOf(Items.POPPY)) {
             change(world, pos, state);
             if (!player.getAbilities().creativeMode) {
                 itemStack.decrement(1);
-                return ActionResult.success(world.isClient);
+                return ItemActionResult.success(world.isClient);
             }
         }
-        return ActionResult.CONSUME;
-    }
-
-    private static boolean isPoppy(ItemStack stack) {
-        return stack.isOf(Items.POPPY);
+        return ItemActionResult.CONSUME;
     }
 
     public static void change(World world, BlockPos pos, BlockState state) {