+++ src/main/java/com/brand/blockus/data/providers/BlockusBlockTagProvider.java
@@ -1457,6 +1457,9 @@ public class BlockusBlockTagProvider extends FabricTagProvider.BlockTagProvider 
         this.getOrCreateTagBuilder(BlockTags.WOOL_CARPETS)
             .addTag(BlockusBlockTags.PATTERNED_CARPETS);
 
+        this.getOrCreateTagBuilder(BlockTags.INSIDE_STEP_SOUND_BLOCKS)
+            .add(RAINBOW_PETAL);
+
         this.getOrCreateTagBuilder(BlockTags.ANIMALS_SPAWNABLE_ON)
             .add(LEGACY_GRASS_BLOCK)
             .add(LEGACY_FIRST_GRASS_BLOCK);