+++ changelog.md
@@ -1,4 +1,12 @@
-* Changed Wooden Lattices & Grates recipes
-  * Lattices are now crafted using 6 grates (Give 16)
-  * Grates are now crafted using 4 sticks around 1 Planks (Give 3)
-* Update zh translations by SolidBlock
+* Updated for 1.21.1 (Compatible with 1.21)
+* Changed Rainbow Rose & Rainbow Petals textures
+* Rainbow Petals can now be placed in the same way as Pink Petals.
+  * Can be bonemealed
+* Changed Rainbow Asphalt, Rainbow Glass & Rainbow Beveled Glass textures
+* Improved Rainbow Asphalt placement.
+* Rainbow Block can now be placed in every direction
+  * Now use "facing" state instead of "axis".
+  * Animation direction now changes depending on facing.
+* Fixed interaction with Legacy Stonecutter, Cookie Block & Weight Storage Cube
+* Fixed Rainbow Glass and Rainbow Glass Pane not being in Creative Inventory.
+